refactor: 계산기에 리뷰 반영
- float을 double 로 변경
- 잘못된 연산자에 대한 에러 처리는 정적 펙터리 메서드로 옮김
- Operand를 사용하는 대신 BiFunction 을 사용하도록 변경

+++ src/test/java/study/calculator/CalculatorTest.java
@@ -22,7 +22,7 @@ public class CalculatorTest {
     @DisplayName("정상적인 식을 테스트 한다.")
     @ParameterizedTest
     @CsvSource(value = {"1 + 1=2", "3 - 1=2", "6 / 3=2", "3 * 7=21", "2 + 3 * 4 / 2=10"}, delimiter = '=')
-    void validFormula(String formula, float expect) {
+    void validFormula(String formula, double expect) {
         assertThat(calculator.calculate(formula)).isEqualTo(expect);
     }
